+++ store-no-op/store.test.ts
@@ -1,17 +1,17 @@
 import {
   open,
-  testIsWriteable,
+  testIsWritable,
   testListItems,
   testSetItem,
   testUrl,
-} from "../store-common/test-storage-module.ts";
+} from "@jollytoad/store-common/test-storage-module";
 import * as store from "./mod.ts";
 
 Deno.test("store-no-op", async (t) => {
   try {
     await open(t, store);
     await testUrl(t, store, "store-no-op");
-    await testIsWriteable(t, store, false);
+    await testIsWritable(t, store, false);
     await testSetItem(t, store);
     await testListItems(t, store, []);
   } finally {