+++ CHANGELOG.md
@@ -6,6 +6,21 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.7.0]
+
+### Added
+
+- [store-extend] new package providing `extendStore` fn
+
+### Changed
+
+- [store-common/test-storage-module] fix `testIsWriteable` -> `testIsWritable`
+  typo
+
+### Fixed
+
+- some relative imports to bare module specifiers in tests
+
 ## [0.6.0]
 
 ### Added